// frontend/apps/veaiops/src/modules/threshold/features/task-config/ui/components/charts/types.ts
// Copyright 2025 Beijing Volcano Engine Technology Co., Ltd. and/or its affiliates
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      https://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export interface AxisDensityOptions {
  maxXTicks?: number;
  xLabelOptionalAngles?: number[];
  timeRange?: [Date, Date]; // 用户选择的时间范围
}